Reduce the concurrency for trustymail scans
We run the three scans separately because we want to reduce the
concurrency for trustymail scans.  This is to avoid a situation where
DNS queries are too high a rate (more than 1024 packets/second) from
the same ENI.  When this happens the DNS queries in excess of 1024
packets/second are dropped, and hence DNS requests time out.

See this link for more information about this VPC DNS limitation:
https://docs.aws.amazon.com/vpc/latest/userguide/vpc-dns.html#vpc-dns-limits

See this link for an explanation as to why the VPC DNS limitation was
not initially a concern:
https://aws.amazon.com/blogs/compute/announcing-improved-vpc-networking-for-aws-lambda-functions/