Reduce sslyze and pshtt concurrency by twenty percent
@climber-girl noticed a case where a pshtt scan failed because no
traffic was returned.  The most likely culprit is that the NAT gateway
through which all the Lambda traffic is passing is being overly taxed
and dropping connections.  To resolve this, I reduced the concurrency
of the pshtt and sslyze scans by twenty percent.

The trustymail concurrency has already been reduced in #35, and there
is no need to reduce it further here.